+++ src/test/suite/web/index.ts
@@ -65,7 +65,6 @@ import '../common/editing_actions.test';
 import '../common/changelog_action.test';
 // Unit prompt-only tests (pure logic)
 import '../../unit-test/delete_lines.simple.test';
-import '../../unit-test/delete_text.simple.test';
 import '../../unit-test/file_actions.simple.test';
 import '../../unit-test/find_text.simple.test';
 import '../../unit-test/get_cursor.simple.test';
@@ -75,7 +74,6 @@ import '../../unit-test/insert_lines.simple.test';
 import '../../unit-test/insert_text.simple.test';
 import '../../unit-test/lint_problems.simple.test';
 import '../../unit-test/place_cursor.simple.test';
-import '../../unit-test/replace_text.simple.test';
 import '../../unit-test/rewrite_all.simple.test';
 import '../../unit-test/rewrite_lines.simple.test';
 import '../../unit-test/tasks.simple.test';