+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Changes between each version before then will not be listed.
   - `delete_file_or_folder` -> `delete_item`
   - `create_file` and `create_folder` have been combined into `create_item`
     - if permissions for both old actions are set, `create_file`'s permission will be the permission set for `create_item` when migrating.
+- Some actions have been merged:
+  - `replace_text` + `delete_text` + `find_text` = `find_and_replace`
 
 ## 2.5.0